Added grouping option + search builder sizing
I added the function to drag one or more columns to the top of the table and remove them with a click while still being able to reorder the columns. I also worked on the search builder field sizing for ipad.  I put new code in for "select type" dropdown allowing it to fill the field properly upon clicking/tabbing.