hiding buttons until results show
I put the code to hide the remove row, download, filter results, and column selection into the ajax so that they will only appear with results. I also added a date filtering widget that may have to be changed because it doesn't work with only a range at the moment.