Two types of filters, column select, images
This version has two types of filters: one in the slider tray and the other in the grid with a field to type.  The tray filter needs work to show counts of items types per column and narrow the selections as filters are chosen.
Image thumbnails are included but need to have links worked in.  Editing button is on each row for testing purposes but probably doesn't need to be on the specimen search page.